+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## [Unreleased]
+### Added
+- Ruby 2.4.1 testing
+
+### Fixed
+- PR template spelling
 
 ## [1.0.0] - 2016-09-25
 ### Changed